fix: share AR request data base
