+++ sglang_omni_v1/scheduling/sglang_backend/request_data.py
@@ -7,24 +7,17 @@ import collections
 from dataclasses import dataclass, field
 from typing import Any
 
+from sglang_omni_v1.scheduling.types import ARRequestData
+
 
 @dataclass
-class SGLangARRequestData:
+class SGLangARRequestData(ARRequestData):
     """Per-request state for SGLang-backed AR stages."""
 
-    input_ids: Any = None
-    attention_mask: Any = None
-    model_inputs: dict[str, Any] = field(default_factory=dict)
-    output_ids: list[int] = field(default_factory=list)
-    extra_model_outputs: dict[str, Any] = field(default_factory=dict)
-    finish_reason: str | None = None
     req: Any = None
     synced: bool = False
     generation_steps: int = 0
     suppress_tokens: list[int] | None = None
-    capture_model_output_keys: tuple = ()
-    max_new_tokens: int | None = None
-    temperature: float = 0.0
     top_p: float = 1.0
     top_k: int = -1
     repetition_penalty: float = 1.0
